requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
boto3==1.36.14
botocore==1.36.14
certifi==2024.8.30
chardet==5.2.0
distro==1.9.0
Django==5.1.2
django-storages==1.14.4
django-widget-tweaks==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
jmespath==1.0.1
openai==1.52.0
pydantic==2.9.2
pydantic_core==2.23.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.3.0
